+++ crc.c
@@ -8,10 +8,10 @@ void buffer_CRC (Bit_stream * bs, unsigned int *old_crc)
   *old_crc = getbits (bs, 16);
 }
 
-void recover_CRC_error (long pcm_sample[7][3][SBLIMIT],
+void recover_CRC_error (int32_t pcm_sample[7][3][SBLIMIT],
 			int error_count,
 			frame_params * fr_ps,
-			FILE * outFile, unsigned long *psampFrames, int ch)
+			FILE * outFile, uint32_t *psampFrames, int ch)
 {
   int stereo = fr_ps->stereo;
   int num, done, i;
